+++ packages/orq-rc/RELEASES.md
@@ -8399,3 +8399,13 @@ Based on:
 - [typescript v4.8.0-rc.32] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.32] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.32 - packages/orq-rc
+
+## 2026-04-20 16:09:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.8 (2.881.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.33] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.33] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.33 - packages/orq-rc